src/requirements.txt
aiogram>=3.13
aiohttp_socks>=0.9
openai>=1.55
httpx[socks]>=0.27
python-dotenv>=1.0
tavily-python>=0.5
requests[socks]>=2.31
yt-dlp>=2024.10
